% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mixgauss.R
\name{create_mixgauss}
\alias{create_mixgauss}
\title{Create length distribution by age}
\usage{
create_mixgauss(age, n, mu, sd)
}
\arguments{
\item{age}{age vector}

\item{n}{number of fishes for each age}

\item{mu}{mean size for each age}

\item{sd}{sd for each age}
}
\value{
tibble with length list
}
\description{
Create length distribution by age
}
\examples{
\dontrun{
age <- c(1, 2)
n   <- c(10, 10)
mu  <- c(30, 50)
sd  <- c(3, 5)
create_mixgauss(age = age, n = n, mu = mu, sd = sd)
}
}
