% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_em.R
\name{fit_em}
\alias{fit_em}
\title{Estimate mixture gaussian distribution by EM algorithm}
\usage{
fit_em(x, mu = c(-1, 1), sigma = c(1, 2), pi = c(0.5, 0.5))
}
\arguments{
\item{x}{distribution}

\item{mu}{mu as vector of initial mean of each gaussian}

\item{sigma}{vector of initial sd of each gaussian}

\item{pi}{vector of sum of initial proportion of mixture gaussian}
}
\description{
Estimate mixture gaussian distribution by EM algorithm
}
