% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate log likelihood}
\usage{
log_likelihood(x, mu, sigma, pi)
}
\arguments{
\item{x}{distribution}

\item{mu}{mean}

\item{sigma}{sd}

\item{pi}{skewness of mixture distribution}
}
\value{
logl ikelihood
}
\description{
Calculate log likelihood
}
