% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixgauss.R
\name{mixgauss}
\alias{mixgauss}
\title{Create length distribution by age}
\usage{
mixgauss(age, n, mu, sd)
}
\arguments{
\item{age}{age vector}

\item{n}{number of fishes for each age}

\item{mu}{mean size for each age}

\item{sd}{sd for each age}
}
\value{
tibble with length list
}
\description{
Create length distribution by age
}
