% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_gaussian.R
\name{mixture_gaussian}
\alias{mixture_gaussian}
\title{Generate random number based on mixture gaussian}
\usage{
mixture_gaussian(x, pi_0 = 0.3, mu_1 = -5, mu_2 = 5, sd_1 = 1, sd_2 = 5)
}
\arguments{
\item{x}{number}

\item{pi_0}{proportion of skewness to gaussian_1. The value must be between 0 <= pi_0 <= 1.}

\item{mu_1}{mean of gaussian_1}

\item{mu_2}{mean of gaussian_2}

\item{sd_1}{sd of gaussian_1}

\item{sd_2}{sd of gaussian_2}
}
\value{

}
\description{
Generate random number based on mixture gaussian
}
